Code Review
Good job on your assignment! All the functionalities on your website works correctly and smoothly, and you covered create, read, update, and delete as the project requires. 
However, when testing the buttons, I noticed that the "All Posts" button on your navbar doesn't take users to a fixed position, and the page shifts down a little bit every time I click it. So maybe you could modify the path for redirecting for this button.